Support recursive type in the LLM.
When recursive type comes in the LLM function calling application schema composer, it had been banned due to `ILlmSchema` does not support the `oneOf` type.

However, this PR makes it possible by supporting a new configurable option, `IHttpLlmApplication.IOptions.recursive`. If you set it to be `false`, the recursive type still be banned. Otherwise you set it to a numeric value, the recursive type would be repeated following the number.